chore - make the runtime-foundation substrate compile and pass its own gate (#1037)
Merges the current dev line, moves the Gate 6/7 additions out of the two monolith files they were sitting in,
and gets the branch from sixty-seven dead-code errors to a clean strict gate.

**Submodules rather than scattered attributes.** The dead code was all in `rustc.rs` (13,042 lines) and
`loaf.rs` (5,042) — both already far past the repository's thousand-line rule, and both full of live code, so a
module-wide allow in either would have hidden future dead code in the live compiler paths. Annotating each item
instead meant fifty-two attributes. So the additions move into three submodules, each carrying one
`#![allow(dead_code, reason = ...)]` that names the gate which consumes it:

    src/oven/rustc/native_input.rs        395 lines   the native-input view: stored plan, candidates, selection
    src/oven/rustc/direct_compiler.rs  1,357 lines   store-owned rustc ownership and the JEC it authorizes
    src/oven/loaf/native_candidates.rs   193 lines   committed-envelope candidate intake

`rustc.rs` drops to 11,336 and `loaf.rs` to 4,878. Two items stay behind with a single-line allow because they
hang off live types — `named_artifacts_for_source_role` on `OvenRustcArtifactManifest`, and
`materialize_declared_rust_libraries` beside the live materializer it will replace. Their justification lives in
the rustdoc, because a multi-line attribute between a doc comment and its item hides the doc from
`check_changed_rustdocs.py`.

**Three failing tests, all fixed, and one of them says something.**

Two failed with `PermissionDenied`: they wrote into a materialized store entry that admission seals read-only.
They now remove and replace the file, which is the pattern the store's own tampering tests already use, and is
also a more honest model of external mutation than writing through the seal.

The third asserted a refusal this branch does not bring. It expected
`project_inspection_test_dependency_envelope_supports_dependencies` to fail with "has no checked Oven source-unit
identity" — a string that appears nowhere in `src/` except that assertion, because it comes from
`work/oven-hot-path`, where Gate 6's blocker has already been implemented. Two things were wrong underneath it:
the fixture's path dependency pointed at a directory that did not exist, so the case failed on "provider artifact
root is not a directory" long before reaching the check; and once the fixture was made real, the function simply
returned `Ok` — this tree reports an unmatched persisted Cargo path root as unsupported rather than refusing it.

The assertion now states what this tree does, and says in a comment that producing the refusal is the Gate 6
blocker and that the blocker replaces this with the negative case. Asserting the refusal here would have been
asserting behaviour the substrate deliberately does not carry.

Gate: `cargo clippy --all-targets --features lsp -- -D warnings` clean, `cargo +nightly fmt --check` clean,
`check_changed_rustdocs.py` passed, 420 `oven::` lib tests pass with none failing.
