+++ src/oven/rustc.rs
@@ -12,10 +12,14 @@ mod inspection;
 // only thing here that needed `rust_inspect`'s selected-projection API — which does not exist on the dev line and
 // belongs in its own change alongside that port. None of Gate 6 or 7's nine review blockers touches it, so leaving
 // it out is what lets the four runtime modules they *do* touch compile. It is retained on `work/oven-hot-path`.
+pub(crate) mod direct_compiler;
+pub(crate) mod native_input;
 mod runtime_closure;
 mod runtime_executor;
 mod runtime_foundation;
 mod selected_unit;
+
+// The native-input view moved into a submodule; callers and this module's own tests still name these directly.
 
 // Split along seams this file already had: the artifact/plan shapes, the project-inspection authority payloads,
 // and the rustc diagnostic report. Every path stays where callers expect it -- re-exported here rather than
@@ -86,386 +90,6 @@ pub const OVEN_RUSTC_REGISTRY_LOCK_RELATIVE_PATH: &str = "registry-sources/Cargo
 pub const OVEN_COMPILER_RUNTIME_CRATE_PREFIX: &str = "incan_";
 /// Schema version for caller-owned native-output reuse evidence.
 const OVEN_DIRECT_RUSTC_OUTPUT_RECEIPT_SCHEMA_VERSION: u32 = 3;
-
-/// Original foundation provenance, preserving distinct store receipts and committed Loaf member identities.
-pub(crate) enum OvenNativeInputOrigin<'facts> {
-    /// A stored direct plan retains the exact original content, receipt and reusable build-unit identities.
-    Store {
-        identity: &'facts str,
-        receipt_identity: &'facts str,
-        build_unit_identity: &'facts str,
-    },
-    /// A toolchain member has committed generation/member/build/plan coordinates, not a store receipt.
-    ToolchainLoaf {
-        generation_identity: &'facts str,
-        member: &'facts super::loaf::OvenLoafEnvelopeMember,
-    },
-}
-
-/// The actual selected owner retained through fact queries and physical attachment.
-enum OvenNativeInputOwner<'owner> {
-    Store(&'owner OvenStoreExecutionPayload),
-    ToolchainLoaf(&'owner super::loaf::OvenMaterializedLoafCandidate<'owner>),
-}
-
-/// Existing stored direct-plan carrier, retaining its original execution owner and one decoded native projection.
-///
-/// Selection keeps the historical inexpensive materialization path. A batch borrows this carrier only after its
-/// existing full owner validation, then enumerates and attaches members without another payload decode or full scan.
-pub(crate) struct OvenStoredDirectRustcExecutionPlan {
-    owner: OvenStoreExecutionPayload,
-    artifacts: OvenRustcArtifactManifest,
-    artifact_plan: OvenRustcArtifactPlan,
-}
-
-impl OvenStoredDirectRustcExecutionPlan {
-    /// Retain an already selected payload and decode its direct plan once, preserving ordinary selection semantics.
-    pub(crate) fn from_execution_payload(owner: OvenStoreExecutionPayload) -> Result<Self, OvenRustcError> {
-        if owner.manifest.kind != OvenArtifactKind::DirectRustcPlan {
-            return Err(OvenRustcError::InvalidInput {
-                field: "stored direct plan",
-                message: "requires an original selected direct-plan payload".to_string(),
-            });
-        }
-        let artifacts: OvenRustcArtifactManifest =
-            serde_json::from_slice(&owner.payload).map_err(|error| OvenRustcError::InvalidInput {
-                field: "stored direct plan payload",
-                message: error.to_string(),
-            })?;
-        let artifact_plan = artifacts.materialize_trusted_store(&owner.artifact_root, &owner.manifest.intent)?;
-        Ok(Self {
-            owner,
-            artifacts,
-            artifact_plan,
-        })
-    }
-
-    /// Return the original immutable Store entry identity.
-    pub(crate) fn identity(&self) -> &str {
-        &self.owner.manifest.identity
-    }
-
-    /// Borrow the sole decoded manifest while its original owner remains retained.
-    pub(crate) fn artifacts(&self) -> &OvenRustcArtifactManifest {
-        &self.artifacts
-    }
-
-    /// Borrow the original materialized root; this is not a caller-supplied authority path.
-    pub(crate) fn artifact_root(&self) -> &Path {
-        &self.owner.artifact_root
-    }
-
-    /// Borrow the existing inexpensive materialized plan used by ordinary stored-plan consumers.
-    pub(crate) fn artifact_plan(&self) -> &OvenRustcArtifactPlan {
-        &self.artifact_plan
-    }
-
-    /// Validate the original owner once for a batch and borrow its retained decode and native projection.
-    pub(crate) fn native_input_view(&self) -> Result<OvenNativeInputView<'_>, OvenRustcError> {
-        self.owner.verify_admitted_payload()?;
-        Ok(OvenNativeInputView {
-            owner: OvenNativeInputOwner::Store(&self.owner),
-            artifacts: Cow::Borrowed(&self.artifacts),
-            artifact_root: self.owner.artifact_root.clone(),
-            plan: Cow::Borrowed(&self.artifact_plan),
-        })
-    }
-}
-
-/// A command-owned physical view of one admitted foundation, borrowing its actual execution owner.
-///
-/// Store construction validates the original record/files once; Loaf construction borrows an already materialized
-/// candidate. Subsequent fact queries do not touch the filesystem or select compatibility. One Incan batch chooses
-/// the foundation before any source unit attaches inputs through it. Neither origin substitutes consumer receipts.
-pub(crate) struct OvenNativeInputView<'owner> {
-    owner: OvenNativeInputOwner<'owner>,
-    artifacts: Cow<'owner, OvenRustcArtifactManifest>,
-    artifact_root: PathBuf,
-    plan: Cow<'owner, OvenRustcArtifactPlan>,
-}
-
-impl<'owner> OvenNativeInputView<'owner> {
-    /// Borrow an original store payload after its existing record and physical integrity checks succeed.
-    pub(crate) fn from_store_payload(owner: &'owner OvenStoreExecutionPayload) -> Result<Self, OvenRustcError> {
-        if owner.manifest.kind != OvenArtifactKind::DirectRustcPlan {
-            return Err(OvenRustcError::InvalidInput {
-                field: "native input owner",
-                message: "requires an original admitted direct-plan payload".to_string(),
-            });
-        }
-        owner.verify_admitted_payload()?;
-        let artifacts: OvenRustcArtifactManifest =
-            serde_json::from_slice(&owner.payload).map_err(|error| OvenRustcError::InvalidInput {
-                field: "native input owner payload",
-                message: error.to_string(),
-            })?;
-        let artifact_root = canonical_directory(&owner.artifact_root, "native input owner root")?;
-        let plan = artifacts.materialize_trusted_store(&artifact_root, &owner.manifest.intent)?;
-        Ok(Self {
-            owner: OvenNativeInputOwner::Store(owner),
-            artifacts: Cow::Owned(artifacts),
-            artifact_root,
-            plan: Cow::Owned(plan),
-        })
-    }
-
-    /// Borrow an already physically validated committed Loaf without repeating its full artifact walk.
-    pub(crate) fn from_materialized_loaf(
-        owner: &'owner super::loaf::OvenMaterializedLoafCandidate<'owner>,
-    ) -> Result<Self, OvenRustcError> {
-        let candidate = owner.candidate();
-        Ok(Self {
-            owner: OvenNativeInputOwner::ToolchainLoaf(owner),
-            artifacts: Cow::Borrowed(&candidate.metadata().plan),
-            artifact_root: canonical_directory(owner.artifact_root(), "native input Loaf root")?,
-            plan: Cow::Borrowed(owner.artifact_plan()),
-        })
-    }
-
-    /// Return the original typed provenance without manufacturing a store receipt for a toolchain member.
-    pub(crate) fn origin(&self) -> OvenNativeInputOrigin<'_> {
-        match &self.owner {
-            OvenNativeInputOwner::Store(owner) => OvenNativeInputOrigin::Store {
-                identity: &owner.manifest.identity,
-                receipt_identity: &owner.manifest.receipt_identity,
-                build_unit_identity: &owner.manifest.build_unit_identity,
-            },
-            OvenNativeInputOwner::ToolchainLoaf(owner) => {
-                let candidate = owner.candidate();
-                OvenNativeInputOrigin::ToolchainLoaf {
-                    generation_identity: candidate.generation_identity(),
-                    member: candidate.member(),
-                }
-            }
-        }
-    }
-
-    /// Return the original admitted content identity, not a new consumer receipt identity.
-    pub(crate) fn identity(&self) -> &str {
-        match self.origin() {
-            OvenNativeInputOrigin::Store { identity, .. } => identity,
-            OvenNativeInputOrigin::ToolchainLoaf { member, .. } => &member.loaf_identity,
-        }
-    }
-
-    /// Return the original store receipt when this origin has one; toolchain members retain different provenance.
-    pub(crate) fn receipt_identity(&self) -> Option<&str> {
-        match self.origin() {
-            OvenNativeInputOrigin::Store { receipt_identity, .. } => Some(receipt_identity),
-            OvenNativeInputOrigin::ToolchainLoaf { .. } => None,
-        }
-    }
-
-    /// Require the selected Store's original full publisher recipe; an exact key cannot supply missing SDK recipe
-    /// facts.
-    pub(crate) fn original_native_receipt(&self) -> Result<&OvenReceipt, OvenRustcError> {
-        let receipt = match &self.owner {
-            OvenNativeInputOwner::Store(owner) => owner.original_native_receipt(),
-            OvenNativeInputOwner::ToolchainLoaf(_) => None,
-        };
-        receipt.ok_or_else(|| OvenRustcError::InvalidInput {
-            field: "native input receipt witness",
-            message: "selected native owner has no original full publisher receipt witness".to_string(),
-        })
-    }
-
-    /// Return the original reusable build-unit identity without recomputing compatibility.
-    pub(crate) fn build_unit_identity(&self) -> &str {
-        match self.origin() {
-            OvenNativeInputOrigin::Store {
-                build_unit_identity, ..
-            } => build_unit_identity,
-            OvenNativeInputOrigin::ToolchainLoaf { member, .. } => &member.build_unit_identity,
-        }
-    }
-
-    /// Borrow the already admitted original catalog for command-table enumeration without another decode or scan.
-    pub(crate) fn artifacts(&self) -> &OvenRustcArtifactManifest {
-        &self.artifacts
-    }
-
-    /// Borrow the admitted physical root only for the existing direct-rustc output containment boundary.
-    pub(crate) fn artifact_root(&self) -> &Path {
-        &self.artifact_root
-    }
-
-    /// Return the original target, toolchain, profile and features as facts, without filtering candidates.
-    pub(crate) fn intent(&self) -> &OvenBuildIntent {
-        &self.artifacts.intent
-    }
-
-    /// Require an original declared source role; a caller's newly generated source key grants nothing here.
-    fn require_original_role(&self, source_role: &str) -> Result<(), OvenRustcError> {
-        if !self.artifacts.entrypoint_externs.contains_key(source_role) {
-            return Err(OvenRustcError::InvalidInput {
-                field: "native input source role",
-                message: format!("`{source_role}` is not an original declared role of this foundation"),
-            });
-        }
-        Ok(())
-    }
-
-    /// Expose every recorded registry leaf in publisher order; version and feature choice belongs to Incan.
-    ///
-    /// A handle retains the original source role for later physical membership checks. Enumeration itself neither
-    /// discovers files nor promotes an anonymous supporting artifact to a named dependency.
-    pub(crate) fn registry_candidates(
-        &self,
-        source_role: &str,
-    ) -> Result<Vec<OvenNativeInputCandidate<'_>>, OvenRustcError> {
-        self.require_original_role(source_role)?;
-        Ok(self
-            .artifacts
-            .registry_leaves
-            .iter()
-            .map(|leaf| OvenNativeInputCandidate {
-                view: self,
-                artifact: &leaf.artifact,
-                registry_leaf: Some(leaf),
-                source_role: source_role.to_string(),
-            })
-            .collect())
-    }
-
-    /// Expose only the explicitly named externs of an original source role, in manifest declaration order.
-    pub(crate) fn named_candidates(
-        &self,
-        source_role: &str,
-    ) -> Result<Vec<OvenNativeInputCandidate<'_>>, OvenRustcError> {
-        self.require_original_role(source_role)?;
-        let projected = self.artifacts.for_source_evidence(source_role)?;
-        Ok(self
-            .artifacts
-            .externs
-            .iter()
-            .filter(|artifact| projected.externs.iter().any(|allowed| allowed == *artifact))
-            .map(|artifact| OvenNativeInputCandidate {
-                view: self,
-                artifact,
-                registry_leaf: None,
-                source_role: source_role.to_string(),
-            })
-            .collect())
-    }
-
-    /// Attach response-selected members transactionally, preserving this exact foundation's original source role.
-    ///
-    /// The host's command table must first bind the Incan response to the global batch and each definition. This
-    /// method checks only original owner/role membership, identifiers, artifact bytes and physical collisions; it
-    /// never interprets dependency requirements. Existing role projection owns search/native paths and environment.
-    /// No missing directory is added, no alternative file is sought and no output or child is created.
-    pub(crate) fn attach_for_source(
-        &self,
-        source_role: &str,
-        selected: &[OvenSelectedNativeInput<'_>],
-    ) -> Result<OvenSelectedNativeSourcePlan<'_>, OvenRustcError> {
-        self.require_original_role(source_role)?;
-        let mut plan = trusted_artifact_plan_for_source_evidence(&self.plan, &self.artifacts, source_role)?;
-        let mut aliases = BTreeSet::new();
-        for input in selected {
-            let candidate = &input.candidate;
-            if !std::ptr::eq(candidate.view, self) || candidate.source_role != source_role {
-                return Err(OvenRustcError::InvalidInput {
-                    field: "selected native input",
-                    message: "belongs to a different foundation view or original source role".to_string(),
-                });
-            }
-            if !aliases.insert(input.alias.as_str()) {
-                return Err(OvenRustcError::InvalidInput {
-                    field: "selected native input",
-                    message: format!("duplicates selected alias `{}`", input.alias),
-                });
-            }
-            let artifact = candidate.artifact;
-            let output = verified_file(
-                &self.artifact_root,
-                &artifact.relative_path,
-                &artifact.digest,
-                "selected native input",
-            )?;
-            let parent = output.parent().ok_or_else(|| OvenRustcError::InvalidInput {
-                field: "selected native input",
-                message: "artifact has no dependency directory".to_string(),
-            })?;
-            if !plan.dependency_search_paths.iter().any(|directory| directory == parent) {
-                return Err(OvenRustcError::InvalidInput {
-                    field: "selected native input",
-                    message: format!(
-                        "{} is outside the original source role's dependency directories",
-                        output.display()
-                    ),
-                });
-            }
-            if let Some((_, existing)) = plan.externs.iter().find(|(name, _)| name == &input.alias) {
-                if existing != &output {
-                    return Err(OvenRustcError::InvalidInput {
-                        field: "selected native input",
-                        message: format!("conflicts with existing named extern `{}`", input.alias),
-                    });
-                }
-                continue;
-            }
-            if plan
-                .caller_owned_library_digests
-                .insert(input.alias.clone(), artifact.digest.clone())
-                .is_some()
-            {
-                return Err(OvenRustcError::InvalidInput {
-                    field: "selected native input",
-                    message: format!("conflicts with existing digest evidence for `{}`", input.alias),
-                });
-            }
-            plan.externs.push((input.alias.clone(), output));
-        }
-        Ok(OvenSelectedNativeSourcePlan { plan, _view: self })
-    }
-}
-
-/// One original manifest member; only its borrowed foundation view can construct this physical handle.
-#[derive(Clone)]
-pub(crate) struct OvenNativeInputCandidate<'view> {
-    view: &'view OvenNativeInputView<'view>,
-    artifact: &'view OvenRustcArtifactExtern,
-    registry_leaf: Option<&'view OvenRustcRegistryLeaf>,
-    source_role: String,
-}
-
-impl<'view> OvenNativeInputCandidate<'view> {
-    /// Return recorded registry source/package/version/features, if this is a registry candidate.
-    pub(crate) fn registry_leaf(&self) -> Option<&OvenRustcRegistryLeaf> {
-        self.registry_leaf
-    }
-
-    /// Return the original named artifact declaration, without resolving its filesystem location.
-    pub(crate) fn artifact(&self) -> &OvenRustcArtifactExtern {
-        self.artifact
-    }
-
-    /// Carry the already selected alias; matching that alias to an authored slot remains the Incan batch decision.
-    pub(crate) fn with_alias(self, alias: String) -> Result<OvenSelectedNativeInput<'view>, OvenRustcError> {
-        validate_rust_identifier(&alias)?;
-        Ok(OvenSelectedNativeInput { candidate: self, alias })
-    }
-}
-
-/// A physical member and explicit direct alias selected by the command's bound Incan batch response.
-pub(crate) struct OvenSelectedNativeInput<'view> {
-    candidate: OvenNativeInputCandidate<'view>,
-    alias: String,
-}
-
-/// A source-specific plan whose actual selected foundation owner remains borrowed through its complete use.
-pub(crate) struct OvenSelectedNativeSourcePlan<'owner> {
-    plan: OvenRustcArtifactPlan,
-    _view: &'owner OvenNativeInputView<'owner>,
-}
-
-impl OvenSelectedNativeSourcePlan<'_> {
-    /// Borrow the attached plan while preserving the selected store lease or committed generation lock.
-    pub(crate) fn artifact_plan(&self) -> &OvenRustcArtifactPlan {
-        &self.plan
-    }
-}
 
 /// One registry leaf and the immutable Loaf root that seals its relative artifact path.
 #[derive(Debug, Clone)]
@@ -1241,179 +865,6 @@ pub struct OvenDirectRustcBake {
     pub reused: bool,
     /// Held for stored consumers until their caller finishes executing the native test binary.
     lease: Option<OvenStoreLease>,
-}
-
-const OVEN_DIRECT_RUSTC_COMPILER_OWNER_SCHEMA_VERSION: u32 = 1;
-const OVEN_DIRECT_RUSTC_COMPILER_DOMAIN_PREFIX: &str = "native-compiler";
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct OvenDirectRustcCompilerOwnerPayload {
-    schema_version: u32,
-    binary_digest: String,
-    closure_digest: String,
-    host: String,
-    target: String,
-    toolchain: String,
-}
-
-/// Toolchain-owned bytes that can affect a direct-Rustc compilation independently of package provenance.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenDirectRustcCompilerEvidence {
-    binary_digest: String,
-    closure_digest: String,
-    host: String,
-    target: String,
-    sysroot: PathBuf,
-    members: Vec<OvenDirectRustcCompilerMember>,
-}
-
-/// One physical compiler member paired with its location-independent sysroot coordinate and byte identity.
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct OvenDirectRustcCompilerMember {
-    relative_path: String,
-    source_path: PathBuf,
-    digest: String,
-}
-
-/// Store-owned compiler closure retained under one lease for an entire native provider batch.
-pub(crate) struct OvenOwnedDirectRustcCompiler {
-    _owner: OvenStoreExecutionPayload,
-    evidence: OvenDirectRustcCompilerEvidence,
-    rustc: PathBuf,
-}
-
-/// Outcome of trying to retain the selected compiler closure for JEC.
-///
-/// Unavailability is deliberately distinct from a hard receipt or intent mismatch: callers may keep compiling
-/// deterministically through their admitted compiler, but must surface why byte-identical reuse was disabled.
-///
-/// `Retained` is boxed because it carries the whole owned compiler, several hundred bytes against `Unavailable`'s
-/// single string. The unavailable arm is the common one on a cold store, and every caller moves the value.
-pub(crate) enum OvenDirectRustcCompilerRetention {
-    Retained(Box<OvenOwnedDirectRustcCompiler>),
-    Unavailable { reason: String },
-}
-
-impl OvenDirectRustcCompilerEvidence {
-    /// Content digest of the `rustc` binary itself, separate from the closure around it.
-    pub(crate) fn binary_digest(&self) -> &str {
-        &self.binary_digest
-    }
-
-    /// Digest over every member of the compiler closure, which is what makes two installations comparable.
-    pub(crate) fn closure_digest(&self) -> &str {
-        &self.closure_digest
-    }
-
-    /// Triple the retained compiler runs on, as opposed to the one it emits for.
-    pub(crate) fn host(&self) -> &str {
-        &self.host
-    }
-
-    /// Triple the retained compiler emits for, as opposed to the one it runs on.
-    pub(crate) fn target(&self) -> &str {
-        &self.target
-    }
-}
-
-impl OvenOwnedDirectRustcCompiler {
-    /// Path to the store-owned `rustc`, valid only while this closure holds its lease.
-    pub(crate) fn rustc(&self) -> &Path {
-        &self.rustc
-    }
-
-    /// Borrow the admitted identity of this retained closure, without exposing the lease that holds it.
-    pub(crate) fn evidence(&self) -> &OvenDirectRustcCompilerEvidence {
-        &self.evidence
-    }
-
-    /// Toolchain identity the closure was admitted under, as the receipt records it.
-    pub(crate) fn toolchain(&self) -> &str {
-        &self._owner.manifest.intent.toolchain
-    }
-}
-
-/// Logical names paired with the exact physical members of one prepared invocation.
-///
-/// The admitted native/source-unit plan decides membership. This binding gives the native-compilation policy stable
-/// names for those already selected paths; it cannot add an extern, search directory or source file.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenDirectRustcJecBindings {
-    pub(crate) logical_source_root: String,
-    pub(crate) logical_entrypoint: String,
-    /// Logical native-member slot for every prepared `--extern`, in exact argument order.
-    pub(crate) externs: Vec<String>,
-    /// Logical search slot for every prepared `-L dependency`, in exact argument order.
-    pub(crate) dependency_searches: Vec<String>,
-    /// Logical search slot for every prepared `-L native`, in exact argument order.
-    pub(crate) native_searches: Vec<String>,
-}
-
-/// A fully observed direct-Rustc library invocation before output lookup or compiler execution.
-///
-/// This owns the source-specific physical plan so the Incan projection and Rustc consume one preparation. Package
-/// name/version and Store coordinates remain outside this value; callers retain them only in request bindings.
-pub(crate) struct OvenPreparedDirectRustcLibrary {
-    rustc: PathBuf,
-    compiler: Option<OvenDirectRustcCompilerEvidence>,
-    /// Lease-protected Store owner for `compiler`; production JEC preparations retain it instead of rereading its
-    /// immutable bytes before every provider unit.
-    compiler_owner: Option<Arc<OvenOwnedDirectRustcCompiler>>,
-    source: PathBuf,
-    source_root: PathBuf,
-    source_digest: String,
-    artifact_root: PathBuf,
-    selected_artifacts: OvenRustcArtifactManifest,
-    plan: OvenRustcArtifactPlan,
-    target: String,
-    profile: String,
-    crate_name: String,
-    edition: String,
-    features: Vec<String>,
-    /// Complete deterministic environment derived from the admitted native-plan values.
-    ///
-    /// Values remain process-local. Callers expose only their digests to the Incan projection.
-    frozen_environment: Option<BTreeMap<String, PathBuf>>,
-}
-
-/// One environment name observed by Rustc, with a digest only when it was present in the frozen environment.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenRustcObservedEnvironment {
-    pub(crate) name: String,
-    pub(crate) value_digest: Option<String>,
-}
-
-/// Parsed Rustc dependency observations. Raw dep-info bytes and environment values never cross this boundary.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) struct OvenRustcDepInfoObservation {
-    pub(crate) files: Vec<PathBuf>,
-    pub(crate) environment: Vec<OvenRustcObservedEnvironment>,
-}
-
-/// Sanitized observation result. An unsupported dep-info shape disables reuse without failing a successful compile.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum OvenRustcDepInfoOutcome {
-    Observed(OvenRustcDepInfoObservation),
-    Unavailable { reason: String },
-}
-
-/// Fresh direct-Rustc output plus the sanitized observations needed for an Incan-owned compilation key.
-pub(crate) struct OvenDirectRustcJecCompilation {
-    pub(crate) bake: OvenDirectRustcBake,
-    pub(crate) observation: OvenRustcDepInfoOutcome,
-}
-
-/// One immutable physical invocation, including its exact output path, argv and cwd.
-///
-/// Cacheable execution uses the prepared frozen environment. An explicitly uncacheable fallback retains the legacy
-/// direct-Rustc inheritance contract.
-pub(crate) struct OvenBoundDirectRustcLibrary<'prepared> {
-    prepared: &'prepared OvenPreparedDirectRustcLibrary,
-    output: PathBuf,
-    arguments: Vec<OsString>,
-    logical_arguments: Vec<Vec<String>>,
-    path_effects_digest: Option<String>,
 }
 
 /// The concrete caller-owned output Rustc must produce for one Oven materialization step.
@@ -3294,6 +2745,10 @@ impl OvenRustcArtifactManifest {
     ///
     /// The existing projection validates the declaration once. Returned records still require an original leased
     /// owner and physical attachment before execution; absent roles cannot use legacy broad projection.
+    ///
+    /// Gate 6 of RFC 119 is the reader. Unlike the rest of this substrate it is annotated rather than moved into a
+    /// submodule, because it hangs off `OvenRustcArtifactManifest`, which is live.
+    #[allow(dead_code, reason = "Gate 6 of RFC 119 is the reader; see the note above")]
     pub(crate) fn named_artifacts_for_source_role(
         &self,
         source_role: &str,
@@ -4652,656 +4107,6 @@ pub fn bake_direct_rustc_test(request: &OvenDirectRustcTestRequest) -> Result<Ov
     )
 }
 
-/// Prepare one provider-owned Rust library from the exact admitted source role and physical native plan.
-///
-/// The result is deliberately output-agnostic: Incan first projects its candidate identity, after which the caller
-/// chooses the corresponding candidate directory for either retained-output validation or a fresh compile.
-pub(crate) fn prepare_trusted_direct_rustc_library_with_artifact_role(
-    request: &OvenTrustedDirectRustcTargetRequest<'_>,
-    artifact_role: &str,
-    source_root: &Path,
-    compiler: Option<&Arc<OvenOwnedDirectRustcCompiler>>,
-) -> Result<OvenPreparedDirectRustcLibrary, OvenRustcError> {
-    request
-        .receipt
-        .verify_identity()
-        .map_err(|error| OvenRustcError::InvalidInput {
-            field: "receipt",
-            message: error.to_string(),
-        })?;
-    if compiler.is_none() {
-        verify_rustc_identity(request.rustc, &request.receipt.intent.toolchain)?;
-    }
-    validate_rust_identifier(request.crate_name)?;
-    validate_edition(request.edition)?;
-    if request.prefer_dynamic {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC library",
-            message: "does not support a dynamic Rust standard-library binding".to_string(),
-        });
-    }
-    if let Some(compiler) = compiler {
-        let evidence = compiler.evidence();
-        let request_rustc = fs::canonicalize(request.rustc).map_err(|source| OvenRustcError::Io {
-            path: request.rustc.to_path_buf(),
-            source,
-        })?;
-        if request_rustc != compiler.rustc
-            || evidence.target != request.receipt.intent.target
-            || compiler.toolchain() != request.receipt.intent.toolchain
-        {
-            return Err(OvenRustcError::InvalidInput {
-                field: "JEC compiler evidence",
-                message: "is detached from the retained receipt-selected compiler owner".to_string(),
-            });
-        }
-    }
-    let source_root = canonical_directory(source_root, "JEC source root")?;
-    let source = fs::canonicalize(verified_regular_file(request.source, "source")?).map_err(|source_error| {
-        OvenRustcError::Io {
-            path: request.source.to_path_buf(),
-            source: source_error,
-        }
-    })?;
-    if !source.starts_with(&source_root) {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC source root",
-            message: "does not contain the prepared source entrypoint".to_string(),
-        });
-    }
-    let source_digest = digest_regular_file(&source, "source")?;
-    let expected_source_digest = request
-        .receipt
-        .sources
-        .supplemental_digests
-        .get(request.source_evidence_key.trim())
-        .ok_or_else(|| OvenRustcError::InvalidInput {
-            field: "source evidence",
-            message: format!("receipt does not declare `{}`", request.source_evidence_key),
-        })?;
-    if expected_source_digest != &source_digest {
-        return Err(OvenRustcError::SourceEvidenceMismatch {
-            key: request.source_evidence_key.to_string(),
-            expected: expected_source_digest.clone(),
-            actual: source_digest,
-        });
-    }
-    let selected_artifacts = request.artifacts.for_source_evidence(artifact_role)?;
-    let plan = if let Some(plan) = request.artifact_plan {
-        trusted_artifact_plan_for_source(plan, request.artifacts, &selected_artifacts)
-    } else {
-        selected_artifacts.materialize_trusted_store(request.artifact_root, &request.receipt.intent)?
-    };
-    let mut compile_environment = BTreeMap::new();
-    for (name, value) in &plan.compile_environment {
-        compile_environment.insert(name.clone(), resolve_compile_environment_value(name, value, &source)?);
-    }
-    let frozen_environment = freeze_direct_rustc_environment(env::vars_os(), &compile_environment);
-    if request.features.windows(2).any(|pair| pair[0] >= pair[1]) {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC features",
-            message: "must be sorted and unique".to_string(),
-        });
-    }
-    Ok(OvenPreparedDirectRustcLibrary {
-        rustc: fs::canonicalize(request.rustc).map_err(|source_error| OvenRustcError::Io {
-            path: request.rustc.to_path_buf(),
-            source: source_error,
-        })?,
-        compiler: compiler.map(|compiler| compiler.evidence.clone()),
-        compiler_owner: compiler.cloned(),
-        source,
-        source_root,
-        source_digest,
-        artifact_root: canonical_directory(request.artifact_root, "artifact root")?,
-        selected_artifacts,
-        plan,
-        target: request.receipt.intent.target.clone(),
-        profile: request.receipt.intent.profile.clone(),
-        crate_name: request.crate_name.to_string(),
-        edition: request.edition.to_string(),
-        features: request.features.to_vec(),
-        frozen_environment,
-    })
-}
-
-impl OvenPreparedDirectRustcLibrary {
-    /// Canonical `rustc` this preparation will launch.
-    pub(crate) fn rustc(&self) -> &Path {
-        &self.rustc
-    }
-
-    /// Admitted compiler identity, absent when this preparation runs without a retained closure.
-    pub(crate) fn compiler(&self) -> Option<&OvenDirectRustcCompilerEvidence> {
-        self.compiler.as_ref()
-    }
-
-    /// Root module this library compiles, already resolved and digest-checked.
-    pub(crate) fn source(&self) -> &Path {
-        &self.source
-    }
-
-    /// Admitted source tree the root module sits in; nothing outside it may reach the compiler.
-    pub(crate) fn source_root(&self) -> &Path {
-        &self.source_root
-    }
-
-    /// Digest of that source tree as it was admitted, so a later step can prove it did not move underneath.
-    pub(crate) fn source_digest(&self) -> &str {
-        &self.source_digest
-    }
-
-    /// Sealed artifact manifest this preparation links against.
-    pub(crate) fn selected_artifacts(&self) -> &OvenRustcArtifactManifest {
-        &self.selected_artifacts
-    }
-
-    /// Store-owned root the sealed artifacts were materialized under.
-    pub(crate) fn artifact_root(&self) -> &Path {
-        &self.artifact_root
-    }
-
-    /// Verified compiler inputs — search paths, externs and environment — from the sealed plan.
-    pub(crate) fn artifact_plan(&self) -> &OvenRustcArtifactPlan {
-        &self.plan
-    }
-
-    /// Target triple from the receipt this preparation was bound to.
-    pub(crate) fn target(&self) -> &str {
-        &self.target
-    }
-
-    /// Named build profile, which decides the codegen flags rather than carrying them.
-    pub(crate) fn profile(&self) -> &str {
-        &self.profile
-    }
-
-    /// Crate name the compiler is told, which is a declared fact rather than one inferred from a path.
-    pub(crate) fn crate_name(&self) -> &str {
-        &self.crate_name
-    }
-
-    /// Rust edition this library is compiled under.
-    pub(crate) fn edition(&self) -> &str {
-        &self.edition
-    }
-
-    /// Resolved feature set, already unified; this is not the declared request.
-    pub(crate) fn features(&self) -> &[String] {
-        &self.features
-    }
-
-    /// Return the complete frozen environment for digest-only host projection.
-    pub(crate) fn environment(&self) -> Option<&BTreeMap<String, PathBuf>> {
-        self.frozen_environment.as_ref()
-    }
-
-    /// Freeze the complete typed projection and the exact physical argv for one output location.
-    pub(crate) fn bind<'prepared>(
-        &'prepared self,
-        output: &Path,
-        bindings: &OvenDirectRustcJecBindings,
-    ) -> Result<OvenBoundDirectRustcLibrary<'prepared>, OvenRustcError> {
-        self.verify_current_inputs()?;
-        self.validate_bindings(bindings)?;
-        let output = caller_output_path(output, &self.artifact_root)?;
-        let mut arguments = Vec::new();
-        let mut logical_arguments = vec![
-            vec!["option".to_string(), "crate-name".to_string(), self.crate_name.clone()],
-            vec!["option".to_string(), "crate-type".to_string(), "rlib".to_string()],
-            vec!["option".to_string(), "edition".to_string(), self.edition.clone()],
-            vec!["option".to_string(), "target".to_string(), self.target.clone()],
-        ];
-        arguments.extend([
-            OsString::from("--crate-name"),
-            OsString::from(&self.crate_name),
-            OsString::from("--crate-type"),
-            OsString::from("rlib"),
-            OsString::from(format!("--edition={}", self.edition)),
-            OsString::from("--target"),
-            OsString::from(&self.target),
-        ]);
-        for (name, value) in oven_profile_codegen_options(&self.profile) {
-            logical_arguments.push(vec!["option".to_string(), name.to_string(), value.to_string()]);
-            arguments.extend([OsString::from("-C"), OsString::from(format!("{name}={value}"))]);
-        }
-        for feature in &self.features {
-            let value = format!("feature={feature:?}");
-            logical_arguments.push(vec!["option".to_string(), "cfg".to_string(), value.clone()]);
-            arguments.extend([OsString::from("--cfg"), OsString::from(value)]);
-        }
-        logical_arguments.extend([
-            vec!["switch".to_string(), "dep-info-stdout".to_string()],
-            vec!["switch".to_string(), "json-diagnostics".to_string()],
-            vec!["remap-source".to_string(), bindings.logical_source_root.clone()],
-        ]);
-        arguments.extend([
-            OsString::from("--error-format=json"),
-            OsString::from("--emit=dep-info=-,link"),
-            joined_os_argument(
-                "--remap-path-prefix=",
-                &self.source_root,
-                Some(&bindings.logical_source_root),
-            ),
-        ]);
-        if let Some(compiler) = &self.compiler {
-            logical_arguments.push(vec!["remap-toolchain".to_string(), "incan-toolchain".to_string()]);
-            arguments.push(joined_os_argument(
-                "--remap-path-prefix=",
-                &compiler.sysroot,
-                Some("incan-toolchain"),
-            ));
-        }
-        for (path, logical) in self
-            .plan
-            .dependency_search_paths
-            .iter()
-            .zip(&bindings.dependency_searches)
-        {
-            logical_arguments.push(vec!["remap-search".to_string(), logical.clone()]);
-            arguments.push(joined_os_argument(
-                "--remap-path-prefix=",
-                path,
-                Some(&format!("incan-native/{logical}")),
-            ));
-            logical_arguments.push(vec!["search".to_string(), logical.clone()]);
-            arguments.extend([OsString::from("-L"), joined_os_argument("dependency=", path, None)]);
-        }
-        for (path, logical) in self.plan.native_search_paths.iter().zip(&bindings.native_searches) {
-            logical_arguments.push(vec!["remap-search".to_string(), logical.clone()]);
-            arguments.push(joined_os_argument(
-                "--remap-path-prefix=",
-                path,
-                Some(&format!("incan-native/{logical}")),
-            ));
-            logical_arguments.push(vec!["search".to_string(), logical.clone()]);
-            arguments.extend([OsString::from("-L"), joined_os_argument("native=", path, None)]);
-        }
-        for ((crate_name, path), logical) in self.plan.externs.iter().zip(&bindings.externs) {
-            logical_arguments.push(vec!["extern".to_string(), logical.clone()]);
-            arguments.extend([
-                OsString::from("--extern"),
-                joined_os_argument(&format!("{crate_name}="), path, None),
-            ]);
-        }
-        logical_arguments.extend([
-            vec!["source".to_string(), bindings.logical_entrypoint.clone()],
-            vec!["output".to_string(), "native".to_string()],
-        ]);
-        arguments.extend([
-            self.source.as_os_str().to_os_string(),
-            OsString::from("-o"),
-            output.as_os_str().to_os_string(),
-        ]);
-        let path_effects_digest = self
-            .compiler
-            .as_ref()
-            .map(|compiler| {
-                serde_json::to_vec(&(
-                    "incan.oven.checked-logical-path-effects/1",
-                    compiler.closure_digest(),
-                    &logical_arguments,
-                ))
-                .map(|material| digest_bytes(&material))
-            })
-            .transpose()
-            .map_err(|error| OvenRustcError::InvalidInput {
-                field: "JEC logical invocation",
-                message: format!("cannot encode path effects: {error}"),
-            })?;
-        Ok(OvenBoundDirectRustcLibrary {
-            prepared: self,
-            output,
-            arguments,
-            logical_arguments,
-            path_effects_digest,
-        })
-    }
-
-    /// Re-check the compiler and sources against what was admitted, immediately before launching.
-    ///
-    /// Preparation validated them once; this runs again at use because nothing holds the filesystem still in
-    /// between. The check is skipped when a store-retained closure owns the compiler, because its lease already
-    /// does hold it, and re-digesting a multi-hundred-megabyte closure per invocation is the cost that lease
-    /// exists to avoid.
-    fn verify_current_inputs(&self) -> Result<(), OvenRustcError> {
-        if self.compiler_owner.is_none() {
-            let compiler = digest_regular_file(&self.rustc, "rustc")?;
-            if let Some(expected) = &self.compiler
-                && compiler != expected.binary_digest
-            {
-                return Err(OvenRustcError::ArtifactDigestMismatch {
-                    path: self.rustc.clone(),
-                    expected: expected.binary_digest.clone(),
-                    actual: compiler,
-                });
-            }
-        } else if self
-            .compiler_owner
-            .as_ref()
-            .is_none_or(|owner| owner.rustc != self.rustc || Some(owner.evidence()) != self.compiler.as_ref())
-        {
-            return Err(OvenRustcError::InvalidInput {
-                field: "JEC compiler owner",
-                message: "is detached from the prepared compiler evidence".to_string(),
-            });
-        }
-        let source = digest_regular_file(&self.source, "source")?;
-        if source != self.source_digest {
-            return Err(OvenRustcError::SourceEvidenceMismatch {
-                key: "prepared JEC source".to_string(),
-                expected: self.source_digest.clone(),
-                actual: source,
-            });
-        }
-        Ok(())
-    }
-
-    /// Refuse logical bindings that could not describe this compilation on another machine.
-    ///
-    /// The bindings are what make a recorded invocation comparable across hosts, so an empty or non-logical one is
-    /// rejected here rather than producing a record that looks portable and is not.
-    fn validate_bindings(&self, bindings: &OvenDirectRustcJecBindings) -> Result<(), OvenRustcError> {
-        let invalid = |message: &str| OvenRustcError::InvalidInput {
-            field: "JEC logical bindings",
-            message: message.to_string(),
-        };
-        if bindings.logical_source_root.trim().is_empty()
-            || bindings.logical_entrypoint.trim().is_empty()
-            || bindings.externs.len() != self.plan.externs.len()
-            || bindings.dependency_searches.len() != self.plan.dependency_search_paths.len()
-            || bindings.native_searches.len() != self.plan.native_search_paths.len()
-        {
-            return Err(invalid("do not cover the prepared invocation"));
-        }
-        if bindings
-            .externs
-            .iter()
-            .chain(&bindings.dependency_searches)
-            .chain(&bindings.native_searches)
-            .any(String::is_empty)
-        {
-            return Err(invalid("substitute or omit a prepared physical member"));
-        }
-        Ok(())
-    }
-}
-
-impl OvenBoundDirectRustcLibrary<'_> {
-    /// Path the compilation wrote, owned by the caller that asked for it.
-    pub(crate) fn output(&self) -> &Path {
-        &self.output
-    }
-
-    /// Compiler invocations as location-independent argument vectors, for comparison across machines.
-    pub(crate) fn logical_arguments(&self) -> &[Vec<String>] {
-        &self.logical_arguments
-    }
-
-    /// Digest of the observed path effects, absent when the compilation ran without observation.
-    pub(crate) fn path_effects_digest(&self) -> Option<&str> {
-        self.path_effects_digest.as_deref()
-    }
-
-    /// Rehash one retained output after the current invocation and Incan key have both been admitted.
-    pub(crate) fn reuse_from(
-        &self,
-        retained_output: &Path,
-        expected_digest: &str,
-    ) -> Result<Option<OvenDirectRustcBake>, OvenRustcError> {
-        let metadata = match fs::symlink_metadata(retained_output) {
-            Err(error) if error.kind() == io::ErrorKind::NotFound => return Ok(None),
-            Err(error) => {
-                return Err(OvenRustcError::Io {
-                    path: retained_output.to_path_buf(),
-                    source: error,
-                });
-            }
-            Ok(metadata) => metadata,
-        };
-        if !metadata.is_file() || metadata.file_type().is_symlink() {
-            return Ok(None);
-        }
-        let output_digest = digest_regular_file(retained_output, "retained output")?;
-        if output_digest != expected_digest {
-            return Ok(None);
-        }
-        let parent = self.output.parent().ok_or_else(|| OvenRustcError::InvalidInput {
-            field: "output",
-            message: "must have a parent directory".to_string(),
-        })?;
-        fs::create_dir_all(parent).map_err(|source| OvenRustcError::Io {
-            path: parent.to_path_buf(),
-            source,
-        })?;
-        let mut temporary = tempfile::NamedTempFile::new_in(parent).map_err(|source| OvenRustcError::Io {
-            path: parent.to_path_buf(),
-            source,
-        })?;
-        let mut retained = fs::File::open(retained_output).map_err(|source| OvenRustcError::Io {
-            path: retained_output.to_path_buf(),
-            source,
-        })?;
-        io::copy(&mut retained, &mut temporary).map_err(|source| OvenRustcError::Io {
-            path: retained_output.to_path_buf(),
-            source,
-        })?;
-        temporary.as_file().sync_all().map_err(|source| OvenRustcError::Io {
-            path: temporary.path().to_path_buf(),
-            source,
-        })?;
-        if digest_regular_file(temporary.path(), "copied retained output")? != expected_digest {
-            return Ok(None);
-        }
-        temporary.persist(&self.output).map_err(|error| OvenRustcError::Io {
-            path: self.output.clone(),
-            source: error.error,
-        })?;
-        Ok(Some(OvenDirectRustcBake {
-            source_digest: self.prepared.source_digest.clone(),
-            output: self.output.clone(),
-            output_digest,
-            cargo_process_started: false,
-            reused: true,
-            lease: None,
-        }))
-    }
-
-    /// Execute the exact frozen argv under its captured environment and return only sanitized dep-info observations.
-    pub(crate) fn compile(&self) -> Result<OvenDirectRustcJecCompilation, OvenRustcError> {
-        self.compile_with_observation(true)
-    }
-
-    /// Execute an explicitly uncacheable invocation under the same compiler environment as a cacheable one.
-    ///
-    /// Cache availability cannot select program semantics. This path suppresses reusable observations, but it still
-    /// clears ambient state and applies only the admitted plan values used by [`Self::compile`].
-    pub(crate) fn compile_uncacheable(&self) -> Result<OvenDirectRustcJecCompilation, OvenRustcError> {
-        self.compile_with_observation(false)
-    }
-
-    /// Run the prepared invocation, optionally recording the paths it touched.
-    ///
-    /// `observe` is a parameter rather than two functions because the compile is identical either way; only
-    /// whether the effects are captured differs, and forking the launch would let the two drift.
-    fn compile_with_observation(&self, observe: bool) -> Result<OvenDirectRustcJecCompilation, OvenRustcError> {
-        self.prepared.verify_current_inputs()?;
-        let mut command = Command::new(&self.prepared.rustc);
-        apply_direct_rustc_execution_environment(&mut command, self.prepared.frozen_environment.as_ref())?;
-        command.current_dir(&self.prepared.source_root).args(&self.arguments);
-        let result = command.output().map_err(|source| OvenRustcError::Io {
-            path: self.prepared.rustc.clone(),
-            source,
-        })?;
-        if !result.status.success() {
-            return Err(OvenRustcError::CompilationFailed {
-                report: parse_rustc_diagnostics(&[], &result.stderr).with_invocation(&command),
-            });
-        }
-        let observation = if observe {
-            let frozen = self
-                .prepared
-                .frozen_environment
-                .as_ref()
-                .ok_or_else(|| OvenRustcError::InvalidInput {
-                    field: "JEC environment",
-                    message: "is unavailable for reusable compilation".to_string(),
-                })?;
-            match parse_rustc_dep_info(&result.stdout, frozen) {
-                Ok(observation) => OvenRustcDepInfoOutcome::Observed(observation),
-                Err(error) => OvenRustcDepInfoOutcome::Unavailable {
-                    reason: error.to_string(),
-                },
-            }
-        } else {
-            OvenRustcDepInfoOutcome::Unavailable {
-                reason: "reuse observation was deliberately suppressed".to_string(),
-            }
-        };
-        verified_regular_file(&self.output, "output")?;
-        let output_digest = digest_regular_file(&self.output, "output")?;
-        Ok(OvenDirectRustcJecCompilation {
-            bake: OvenDirectRustcBake {
-                source_digest: self.prepared.source_digest.clone(),
-                output: self.output.clone(),
-                output_digest,
-                cargo_process_started: false,
-                reused: false,
-                lease: None,
-            },
-            observation,
-        })
-    }
-}
-
-/// Apply the one deterministic direct-Rustc environment used with or without cache acceleration.
-fn apply_direct_rustc_execution_environment(
-    command: &mut Command,
-    frozen_environment: Option<&BTreeMap<String, PathBuf>>,
-) -> Result<(), OvenRustcError> {
-    let frozen_environment = frozen_environment.ok_or_else(|| OvenRustcError::InvalidInput {
-        field: "JEC environment",
-        message: "is unavailable for direct compilation".to_string(),
-    })?;
-    command.env_clear().envs(frozen_environment);
-    Ok(())
-}
-
-/// Construct one path-bearing argument without lossy UTF-8 conversion.
-fn joined_os_argument(prefix: &str, path: &Path, suffix: Option<&str>) -> OsString {
-    let mut value = OsString::from(prefix);
-    value.push(path.as_os_str());
-    if let Some(suffix) = suffix {
-        value.push("=");
-        value.push(suffix);
-    }
-    value
-}
-
-/// Parse Rustc's Makefile dep-info and discard every raw environment value before returning.
-fn parse_rustc_dep_info(
-    bytes: &[u8],
-    environment: &BTreeMap<String, PathBuf>,
-) -> Result<OvenRustcDepInfoObservation, OvenRustcError> {
-    let text = std::str::from_utf8(bytes).map_err(|error| OvenRustcError::InvalidInput {
-        field: "JEC dep-info",
-        message: format!("is not UTF-8: {error}"),
-    })?;
-    let unfolded = text.replace("\\\r\n", "").replace("\\\n", "");
-    let dependencies = unfolded
-        .lines()
-        .filter_map(|line| {
-            let line = line.trim_end_matches('\r');
-            (!line.trim().is_empty() && !line.starts_with('#')).then_some(line)
-        })
-        .find_map(|line| line.find(": ").map(|separator| &line[separator + 2..]))
-        .ok_or_else(|| OvenRustcError::InvalidInput {
-            field: "JEC dep-info",
-            message: "has no dependency rule".to_string(),
-        })?;
-    let files = parse_makefile_words(dependencies)?;
-    if files.is_empty() {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC dep-info",
-            message: "has an empty dependency rule".to_string(),
-        });
-    }
-    let mut files = files.into_iter().map(PathBuf::from).collect::<Vec<_>>();
-    files.sort();
-    files.dedup();
-
-    let mut observed = BTreeMap::new();
-    for line in unfolded.lines() {
-        let Some(value) = line.trim_end_matches('\r').strip_prefix("# env-dep:") else {
-            continue;
-        };
-        let (name, supplied) = value
-            .split_once('=')
-            .map_or((value, None), |(name, value)| (name, Some(value)));
-        if name.is_empty() || observed.contains_key(name) {
-            return Err(OvenRustcError::InvalidInput {
-                field: "JEC dep-info",
-                message: "has an empty or repeated environment name".to_string(),
-            });
-        }
-        let current = environment.get(name);
-        let matches = match (supplied, current) {
-            (None, None) => true,
-            (Some(supplied), Some(current)) => current.to_str() == Some(supplied),
-            _ => false,
-        };
-        if !matches {
-            return Err(OvenRustcError::InvalidInput {
-                field: "JEC dep-info",
-                message: format!("does not match the frozen value presence for `{name}`"),
-            });
-        }
-        observed.insert(
-            name.to_string(),
-            current.map(|value| digest_bytes(value.as_os_str().as_encoded_bytes())),
-        );
-    }
-    Ok(OvenRustcDepInfoObservation {
-        files,
-        environment: observed
-            .into_iter()
-            .map(|(name, value_digest)| OvenRustcObservedEnvironment { name, value_digest })
-            .collect(),
-    })
-}
-
-/// Decode the subset of Makefile word escaping emitted by Rustc dep-info.
-fn parse_makefile_words(value: &str) -> Result<Vec<String>, OvenRustcError> {
-    let mut words = Vec::new();
-    let mut word = String::new();
-    let mut characters = value.chars().peekable();
-    while let Some(character) = characters.next() {
-        match character {
-            '\\' => {
-                let escaped = characters.next().ok_or_else(|| OvenRustcError::InvalidInput {
-                    field: "JEC dep-info",
-                    message: "ends with an incomplete path escape".to_string(),
-                })?;
-                word.push(escaped);
-            }
-            '$' if characters.peek() == Some(&'$') => {
-                characters.next();
-                word.push('$');
-            }
-            character if character.is_whitespace() => {
-                if !word.is_empty() {
-                    words.push(std::mem::take(&mut word));
-                }
-            }
-            character => word.push(character),
-        }
-    }
-    if !word.is_empty() {
-        words.push(word);
-    }
-    Ok(words)
-}
-
 /// Compile one compiler-suite test root from a caller-held leased store artifact without a Cargo consumer process.
 pub(crate) fn bake_trusted_direct_rustc_test(
     request: &OvenTrustedDirectRustcTargetRequest<'_>,
@@ -5830,18 +4635,6 @@ fn write_caller_output_record(output: &Path, record: &OvenDirectRustcOutputRecor
     fs::rename(&temporary, &path).map_err(|source| OvenRustcError::Io { path, source })
 }
 
-/// Build the complete reusable compiler environment from admitted plan values only.
-///
-/// Ambient values are intentionally ignored. Rustc controls and dynamic-loader paths can change compiler behavior
-/// without appearing as source `env!` dependencies, so inherited process state cannot participate in a reusable
-/// compilation unless a later policy models it explicitly.
-fn freeze_direct_rustc_environment(
-    _inherited: impl IntoIterator<Item = (OsString, OsString)>,
-    compile_environment: &BTreeMap<String, PathBuf>,
-) -> Option<BTreeMap<String, PathBuf>> {
-    Some(compile_environment.clone())
-}
-
 /// Whether one inherited environment variable must be cleared before launching `rustc`.
 ///
 /// Anything Cargo sets, plus the wrapper and flag variables, changes what the compiler does without appearing in
@@ -6190,511 +4983,6 @@ fn rustc_sysroot(rustc: &Path) -> Result<PathBuf, OvenRustcError> {
         });
     }
     Ok(sysroot)
-}
-
-/// Hash the selected compiler and the bounded sysroot closure that can affect one direct `rlib` compilation.
-///
-/// Package names, versions and installation locations are deliberately absent. The projection contains the invoked
-/// compiler bytes, its actual sysroot compiler when distinct, the driver/LLVM libraries beside that compiler, and
-/// the host/target Rust libraries selected by the invocation. Other installed targets are not members.
-pub(crate) fn direct_rustc_compiler_evidence(
-    rustc: &Path,
-    target: &str,
-) -> Result<OvenDirectRustcCompilerEvidence, OvenRustcError> {
-    const MAX_MEMBERS: usize = 8_192;
-    const MAX_BYTES: u64 = 16 * 1024 * 1024 * 1024;
-
-    if target.trim().is_empty() {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC compiler target",
-            message: "must not be empty".to_string(),
-        });
-    }
-    let rustc = fs::canonicalize(verified_regular_file(rustc, "rustc")?).map_err(|source| OvenRustcError::Io {
-        path: rustc.to_path_buf(),
-        source,
-    })?;
-    let host = rustc_host_target(&rustc)?;
-    let sysroot = fs::canonicalize(rustc_sysroot(&rustc)?).map_err(|source| OvenRustcError::Io {
-        path: rustc.clone(),
-        source,
-    })?;
-    let binary_digest = digest_regular_file(&rustc, "rustc")?;
-    let sysroot_rustc =
-        fs::canonicalize(verified_regular_file(&sysroot.join("bin/rustc"), "sysroot rustc")?).map_err(|source| {
-            OvenRustcError::Io {
-                path: sysroot.join("bin/rustc"),
-                source,
-            }
-        })?;
-    if sysroot_rustc != rustc {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: "requires the invoked compiler to be the selected sysroot's own bin/rustc".to_string(),
-        });
-    }
-    let mut members = BTreeMap::from([("bin/rustc".to_string(), (rustc.clone(), binary_digest.clone()))]);
-    let mut total_bytes = fs::metadata(&rustc)
-        .map_err(|source| OvenRustcError::Io {
-            path: rustc.clone(),
-            source,
-        })?
-        .len();
-
-    collect_compiler_closure_directory(
-        &sysroot.join("lib"),
-        "lib",
-        false,
-        &mut members,
-        &mut total_bytes,
-        MAX_MEMBERS,
-        MAX_BYTES,
-    )?;
-    let mut targets = BTreeSet::from([host.clone(), target.to_string()]);
-    for selected in std::mem::take(&mut targets) {
-        let root = sysroot.join("lib/rustlib").join(&selected);
-        for child in ["lib", "codegen-backends"] {
-            let path = root.join(child);
-            match fs::symlink_metadata(&path) {
-                Ok(metadata) if metadata.is_dir() && !metadata.file_type().is_symlink() => {
-                    collect_compiler_closure_directory(
-                        &path,
-                        &format!("lib/rustlib/{selected}/{child}"),
-                        true,
-                        &mut members,
-                        &mut total_bytes,
-                        MAX_MEMBERS,
-                        MAX_BYTES,
-                    )?;
-                }
-                Err(error) if error.kind() == io::ErrorKind::NotFound && child == "codegen-backends" => {}
-                Err(error) => {
-                    return Err(OvenRustcError::Io { path, source: error });
-                }
-                _ => {
-                    return Err(OvenRustcError::InvalidInput {
-                        field: "JEC compiler closure",
-                        message: format!("{} is not a non-symlink directory", path.display()),
-                    });
-                }
-            }
-        }
-    }
-    let member_digests = members
-        .iter()
-        .map(|(path, (_, digest))| (path.clone(), digest.clone()))
-        .collect::<BTreeMap<_, _>>();
-    let closure_digest = direct_rustc_compiler_closure_digest(&member_digests)?;
-    let members = members
-        .into_iter()
-        .map(|(relative_path, (source_path, digest))| OvenDirectRustcCompilerMember {
-            relative_path,
-            source_path,
-            digest,
-        })
-        .collect();
-    Ok(OvenDirectRustcCompilerEvidence {
-        binary_digest,
-        closure_digest,
-        host,
-        target: target.to_string(),
-        sysroot,
-        members,
-    })
-}
-
-/// Digest one compiler closure from its members' logical coordinates and byte identities.
-///
-/// The schema tag is folded in so a later change to what a closure contains cannot silently collide with an
-/// identity minted under the old shape.
-fn direct_rustc_compiler_closure_digest(members: &BTreeMap<String, String>) -> Result<String, OvenRustcError> {
-    let material = serde_json::to_vec(&("incan.oven.rustc-rlib-closure/1", members)).map_err(|error| {
-        OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: format!("cannot encode compiler-member identities: {error}"),
-        }
-    })?;
-    Ok(digest_bytes(&material))
-}
-
-/// Derive the store compatibility domain one compiler closure is published into.
-///
-/// Keying the domain on the closure digest is what keeps two installations of the same toolchain version from
-/// sharing an entry when their closures actually differ.
-fn direct_rustc_compiler_domain(closure_digest: &str) -> Result<String, OvenRustcError> {
-    let Some(hex) = closure_digest.strip_prefix("sha256:") else {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: "has no SHA-256 identity prefix".to_string(),
-        });
-    };
-    if hex.len() != 64
-        || !hex
-            .bytes()
-            .all(|byte| byte.is_ascii_digit() || (b'a'..=b'f').contains(&byte))
-    {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: "has a malformed SHA-256 identity".to_string(),
-        });
-    }
-    Ok(format!("{OVEN_DIRECT_RUSTC_COMPILER_DOMAIN_PREFIX}.{hex}"))
-}
-
-/// Select or materialize one compiler closure in the immutable Store and retain its lease for native execution.
-///
-/// Discovery, capacity or Store availability failures disable JEC and leave ordinary deterministic direct-Rustc
-/// compilation available. A single compatible Store closure is verified and preferred before the ambient compiler is
-/// executed or its sysroot is hashed. Multiple byte-distinct closures with the same target/toolchain identity force a
-/// cold observation so the ambient compiler bytes disambiguate them.
-///
-/// Once selected, that verified Store closure is Oven's compiler authority for the batch. The caller's mutable Rustc
-/// path supplies identity and bytes only for a cold or ambiguous selection; it is never executed in place when a
-/// unique warm owner can be admitted.
-pub(crate) fn retain_direct_rustc_compiler(
-    store: &OvenStore,
-    receipt: &OvenReceipt,
-    candidate_rustc: &Path,
-    target: &str,
-) -> Result<OvenDirectRustcCompilerRetention, OvenRustcError> {
-    receipt
-        .verify_identity()
-        .map_err(|error| OvenRustcError::InvalidInput {
-            field: "JEC compiler receipt",
-            message: error.to_string(),
-        })?;
-    if receipt.intent.target != target {
-        return Err(OvenRustcError::IntentMismatch);
-    }
-    let toolchain = receipt.intent.toolchain.clone();
-    let mut selected = match store.select_payloads_matching_for_execution(|manifest| {
-        manifest.kind == OvenArtifactKind::NativeCompilerClosure
-            && manifest.intent.target == target
-            && manifest.intent.toolchain == toolchain
-    }) {
-        Ok(mut selected) => {
-            selected.sort_by(|left, right| left.manifest.identity.cmp(&right.manifest.identity));
-            selected
-        }
-        Err(error) => {
-            return Ok(OvenDirectRustcCompilerRetention::Unavailable {
-                reason: format!("cannot inspect retained compiler closures: {error}"),
-            });
-        }
-    };
-    let compatible = selected
-        .iter()
-        .enumerate()
-        .filter_map(|(index, owner)| {
-            matching_direct_rustc_compiler_owner_payload(owner, target, &toolchain, None)
-                .map(|payload| (index, payload))
-        })
-        .collect::<Vec<_>>();
-    let closure_identities = compatible
-        .iter()
-        .map(|(_, payload)| {
-            (
-                payload.closure_digest.as_str(),
-                payload.binary_digest.as_str(),
-                payload.host.as_str(),
-            )
-        })
-        .collect::<BTreeSet<_>>();
-    if closure_identities.len() == 1
-        && let Some((index, payload)) = compatible.first()
-    {
-        let host = payload.host.clone();
-        let owner = selected.remove(*index);
-        if let Some(owned) = admit_direct_rustc_compiler_owner(owner, target, &toolchain, &host, None)? {
-            return Ok(OvenDirectRustcCompilerRetention::Retained(Box::new(owned)));
-        }
-    }
-
-    let candidate_toolchain = rustc_identity(candidate_rustc)?;
-    if toolchain != candidate_toolchain {
-        return Err(OvenRustcError::ToolchainMismatch {
-            expected: toolchain,
-            actual: candidate_toolchain,
-        });
-    }
-    let host = rustc_host_target(candidate_rustc)?;
-    let evidence = match direct_rustc_compiler_evidence(candidate_rustc, target) {
-        Ok(evidence) => evidence,
-        Err(error) => {
-            return Ok(OvenDirectRustcCompilerRetention::Unavailable {
-                reason: format!("cannot admit the selected compiler closure: {error}"),
-            });
-        }
-    };
-    if evidence.host != host {
-        return Ok(OvenDirectRustcCompilerRetention::Unavailable {
-            reason: "the selected compiler changed its reported host during closure observation".to_string(),
-        });
-    }
-    let payload = OvenDirectRustcCompilerOwnerPayload {
-        schema_version: OVEN_DIRECT_RUSTC_COMPILER_OWNER_SCHEMA_VERSION,
-        binary_digest: evidence.binary_digest.clone(),
-        closure_digest: evidence.closure_digest.clone(),
-        host: evidence.host.clone(),
-        target: evidence.target.clone(),
-        toolchain: receipt.intent.toolchain.clone(),
-    };
-    let domain = direct_rustc_compiler_domain(&evidence.closure_digest)?;
-    if let Some(index) = selected.iter().position(|owner| {
-        matching_direct_rustc_compiler_owner_payload(owner, target, &toolchain, Some(&host)).as_ref() == Some(&payload)
-    }) {
-        let owner = selected.remove(index);
-        if let Some(owned) = admit_direct_rustc_compiler_owner(owner, target, &toolchain, &host, Some(&evidence))? {
-            return Ok(OvenDirectRustcCompilerRetention::Retained(Box::new(owned)));
-        }
-    }
-
-    let encoded = serde_json::to_vec(&payload).map_err(|error| OvenRustcError::InvalidInput {
-        field: "JEC compiler owner",
-        message: format!("cannot encode compiler closure metadata: {error}"),
-    })?;
-    let materialized_files = evidence
-        .members
-        .iter()
-        .map(|member| OvenArtifactMaterializedFile {
-            source_path: member.source_path.clone(),
-            relative_path: member.relative_path.clone(),
-        })
-        .collect();
-    let manifest = match store.publish(&OvenArtifactPublishRequest {
-        receipt: receipt.clone(),
-        domain: domain.clone(),
-        kind: OvenArtifactKind::NativeCompilerClosure,
-        payload: encoded,
-        materialized_files,
-    }) {
-        Ok(manifest) => manifest,
-        Err(error) => {
-            return Ok(OvenDirectRustcCompilerRetention::Unavailable {
-                reason: format!("cannot publish the selected compiler closure: {error}"),
-            });
-        }
-    };
-    let mut selected = match store.select_payloads_for_execution(std::slice::from_ref(&manifest.identity)) {
-        Ok(selected) => selected,
-        Err(error) => {
-            return Ok(OvenDirectRustcCompilerRetention::Unavailable {
-                reason: format!("cannot retain the published compiler closure: {error}"),
-            });
-        }
-    };
-    let Some(owner) = selected.pop() else {
-        return Ok(OvenDirectRustcCompilerRetention::Unavailable {
-            reason: "the published compiler closure has no execution owner".to_string(),
-        });
-    };
-    match admit_direct_rustc_compiler_owner(owner, target, &toolchain, &host, Some(&evidence))? {
-        Some(owned) => Ok(OvenDirectRustcCompilerRetention::Retained(Box::new(owned))),
-        None => Ok(OvenDirectRustcCompilerRetention::Unavailable {
-            reason: "the published compiler closure failed final admission".to_string(),
-        }),
-    }
-}
-
-/// Decode the small owner descriptor used to select a warm compiler without reading its materialized closure.
-fn matching_direct_rustc_compiler_owner_payload(
-    owner: &OvenStoreExecutionPayload,
-    target: &str,
-    toolchain: &str,
-    host: Option<&str>,
-) -> Option<OvenDirectRustcCompilerOwnerPayload> {
-    if owner.manifest.kind != OvenArtifactKind::NativeCompilerClosure
-        || owner.manifest.intent.target != target
-        || owner.manifest.intent.toolchain != toolchain
-    {
-        return None;
-    }
-    let payload = serde_json::from_slice::<OvenDirectRustcCompilerOwnerPayload>(&owner.payload).ok()?;
-    if payload.schema_version != OVEN_DIRECT_RUSTC_COMPILER_OWNER_SCHEMA_VERSION
-        || payload.target != target
-        || payload.toolchain != toolchain
-        || host.is_some_and(|host| payload.host != host)
-        || direct_rustc_compiler_domain(&payload.closure_digest).ok().as_deref() != Some(owner.manifest.domain.as_str())
-    {
-        return None;
-    }
-    Some(payload)
-}
-
-/// Admit one retained store entry as this batch's compiler, or say why it cannot serve.
-///
-/// The entry's recorded host, target and toolchain must match the request, and where the caller already holds
-/// expected evidence it must match that too. A mismatch is reported as unavailability rather than as an error:
-/// the caller can still compile through its own admitted compiler, it just loses byte-identical reuse.
-fn admit_direct_rustc_compiler_owner(
-    owner: OvenStoreExecutionPayload,
-    target: &str,
-    toolchain: &str,
-    host: &str,
-    expected: Option<&OvenDirectRustcCompilerEvidence>,
-) -> Result<Option<OvenOwnedDirectRustcCompiler>, OvenRustcError> {
-    let Some(payload) = matching_direct_rustc_compiler_owner_payload(&owner, target, toolchain, Some(host)) else {
-        return Ok(None);
-    };
-    if expected.is_some_and(|expected| {
-        payload.binary_digest != expected.binary_digest
-            || payload.closure_digest != expected.closure_digest
-            || payload.host != expected.host
-            || payload.target != expected.target
-    }) {
-        return Ok(None);
-    }
-    if owner.verify_materialized_files().is_err() {
-        return Ok(None);
-    }
-    let mut member_digests = BTreeMap::new();
-    let mut members = Vec::with_capacity(owner.manifest.materialized_files.len());
-    let mut rustc = None;
-    for member in &owner.manifest.materialized_files {
-        if member_digests
-            .insert(member.relative_path.clone(), member.digest.clone())
-            .is_some()
-        {
-            return Ok(None);
-        }
-        let source_path = owner.artifact_root.join(&member.relative_path);
-        if member.relative_path == "bin/rustc" {
-            if !member.executable || member.digest != payload.binary_digest {
-                return Ok(None);
-            }
-            rustc = Some(source_path.clone());
-        }
-        members.push(OvenDirectRustcCompilerMember {
-            relative_path: member.relative_path.clone(),
-            source_path,
-            digest: member.digest.clone(),
-        });
-    }
-    if direct_rustc_compiler_closure_digest(&member_digests)? != payload.closure_digest {
-        return Ok(None);
-    }
-    let Some(rustc) = rustc else {
-        return Ok(None);
-    };
-    let sysroot = match fs::canonicalize(&owner.artifact_root) {
-        Ok(sysroot) => sysroot,
-        Err(_) => return Ok(None),
-    };
-    if fs::canonicalize(rustc_sysroot(&rustc)?).ok().as_deref() != Some(sysroot.as_path())
-        || rustc_host_target(&rustc)? != payload.host
-        || rustc_identity(&rustc)? != payload.toolchain
-    {
-        return Ok(None);
-    }
-    Ok(Some(OvenOwnedDirectRustcCompiler {
-        _owner: owner,
-        evidence: OvenDirectRustcCompilerEvidence {
-            binary_digest: payload.binary_digest.clone(),
-            closure_digest: payload.closure_digest.clone(),
-            host: payload.host.clone(),
-            target: payload.target.clone(),
-            sysroot,
-            members,
-        },
-        rustc,
-    }))
-}
-
-#[allow(clippy::too_many_arguments)]
-/// Walk one directory of the compiler installation into logical members, bounded by count and bytes.
-///
-/// The bounds are the point: this reads a directory the caller named, and an unbounded walk of a sysroot would
-/// admit an arbitrary amount of material into an identity. `logical_root` keeps each member's coordinate
-/// independent of where the installation happens to live.
-fn collect_compiler_closure_directory(
-    root: &Path,
-    logical_root: &str,
-    recursive: bool,
-    members: &mut BTreeMap<String, (PathBuf, String)>,
-    total_bytes: &mut u64,
-    max_members: usize,
-    max_bytes: u64,
-) -> Result<(), OvenRustcError> {
-    let mut entries = fs::read_dir(root)
-        .map_err(|source| OvenRustcError::Io {
-            path: root.to_path_buf(),
-            source,
-        })?
-        .collect::<Result<Vec<_>, _>>()
-        .map_err(|source| OvenRustcError::Io {
-            path: root.to_path_buf(),
-            source,
-        })?;
-    entries.sort_by_key(|entry| entry.file_name());
-    for entry in entries {
-        let path = entry.path();
-        let name = entry
-            .file_name()
-            .into_string()
-            .map_err(|_| OvenRustcError::InvalidInput {
-                field: "JEC compiler closure",
-                message: "contains a non-UTF-8 member name".to_string(),
-            })?;
-        let logical = format!("{logical_root}/{name}");
-        let metadata = fs::symlink_metadata(&path).map_err(|source| OvenRustcError::Io {
-            path: path.clone(),
-            source,
-        })?;
-        if metadata.file_type().is_symlink() {
-            return Err(OvenRustcError::InvalidInput {
-                field: "JEC compiler closure",
-                message: format!("contains a symlink at {logical}"),
-            });
-        }
-        if metadata.is_file() {
-            collect_compiler_closure_file(&path, logical, members, total_bytes, max_members, max_bytes)?;
-        } else if metadata.is_dir() && recursive {
-            collect_compiler_closure_directory(&path, &logical, true, members, total_bytes, max_members, max_bytes)?;
-        }
-    }
-    Ok(())
-}
-
-/// Admit one compiler-closure file under its logical coordinate, enforcing the member and byte bounds.
-fn collect_compiler_closure_file(
-    path: &Path,
-    logical: String,
-    members: &mut BTreeMap<String, (PathBuf, String)>,
-    total_bytes: &mut u64,
-    max_members: usize,
-    max_bytes: u64,
-) -> Result<(), OvenRustcError> {
-    let metadata = fs::symlink_metadata(path).map_err(|source| OvenRustcError::Io {
-        path: path.to_path_buf(),
-        source,
-    })?;
-    if !metadata.is_file() || metadata.file_type().is_symlink() {
-        return Err(OvenRustcError::InvalidArtifactPath {
-            kind: "JEC compiler closure",
-            path: path.to_path_buf(),
-            message: "must contain only non-symlink regular files".to_string(),
-        });
-    }
-    *total_bytes = total_bytes
-        .checked_add(metadata.len())
-        .ok_or_else(|| OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: "byte count overflowed".to_string(),
-        })?;
-    if members.len() >= max_members || *total_bytes > max_bytes {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: "exceeds its bounded member or byte budget".to_string(),
-        });
-    }
-    let digest = digest_regular_file(path, "JEC compiler closure member")?;
-    if members.insert(logical, (path.to_path_buf(), digest)).is_some() {
-        return Err(OvenRustcError::InvalidInput {
-            field: "JEC compiler closure",
-            message: "contains a duplicate logical member".to_string(),
-        });
-    }
-    Ok(())
 }
 
 /// Resolve the Rustdoc executable from the same verified sysroot as a receipt-selected compiler.
@@ -7445,7 +5733,7 @@ mod tests {
             ("OVERRIDE".to_string(), PathBuf::from("plan")),
         ]);
 
-        let frozen = super::freeze_direct_rustc_environment(inherited, &compile_environment)
+        let frozen = super::direct_compiler::freeze_direct_rustc_environment(inherited, &compile_environment)
             .ok_or("test environment was not representable")?;
 
         assert_eq!(frozen.get("OVERRIDE"), Some(&PathBuf::from("plan")));
@@ -7473,12 +5761,12 @@ mod tests {
 
         let inherited = [(OsString::from_vec(vec![0xff]), OsString::from("value"))];
         assert_eq!(
-            super::freeze_direct_rustc_environment(inherited, &BTreeMap::new()),
+            super::direct_compiler::freeze_direct_rustc_environment(inherited, &BTreeMap::new()),
             Some(BTreeMap::new())
         );
         let mut command = Command::new("/usr/bin/env");
         command.env("AMBIENT_VALUE", "ambient");
-        super::apply_direct_rustc_execution_environment(&mut command, Some(&BTreeMap::new()))?;
+        super::direct_compiler::apply_direct_rustc_execution_environment(&mut command, Some(&BTreeMap::new()))?;
         let output = String::from_utf8(command.output()?.stdout)?;
         assert!(!output.lines().any(|line| line.starts_with("AMBIENT_VALUE=")));
         Ok(())
@@ -7489,7 +5777,7 @@ mod tests {
     fn jec_execution_environment_is_identical_with_or_without_cache_observation()
     -> Result<(), Box<dyn std::error::Error>> {
         let compile_environment = BTreeMap::from([("PLAN_VALUE".to_string(), PathBuf::from("plan"))]);
-        let frozen_environment = super::freeze_direct_rustc_environment(
+        let frozen_environment = super::direct_compiler::freeze_direct_rustc_environment(
             [(OsString::from("FROZEN_VALUE"), OsString::from("frozen"))],
             &compile_environment,
         )
@@ -7497,7 +5785,7 @@ mod tests {
 
         let mut cacheable = Command::new("/usr/bin/env");
         cacheable.env("LEGACY_VALUE", "ambient");
-        super::apply_direct_rustc_execution_environment(&mut cacheable, Some(&frozen_environment))?;
+        super::direct_compiler::apply_direct_rustc_execution_environment(&mut cacheable, Some(&frozen_environment))?;
         let cacheable_output = String::from_utf8(cacheable.output()?.stdout)?;
         assert!(cacheable_output.lines().any(|line| line == "PLAN_VALUE=plan"));
         assert!(!cacheable_output.lines().any(|line| line.starts_with("FROZEN_VALUE=")));
@@ -7505,7 +5793,7 @@ mod tests {
 
         let mut uncacheable = Command::new("/usr/bin/env");
         uncacheable.env("LEGACY_VALUE", "ambient");
-        super::apply_direct_rustc_execution_environment(&mut uncacheable, Some(&frozen_environment))?;
+        super::direct_compiler::apply_direct_rustc_execution_environment(&mut uncacheable, Some(&frozen_environment))?;
         let uncacheable_output = String::from_utf8(uncacheable.output()?.stdout)?;
         assert!(uncacheable_output.lines().any(|line| line == "PLAN_VALUE=plan"));
         assert!(!uncacheable_output.lines().any(|line| line.starts_with("FROZEN_VALUE=")));
@@ -7518,7 +5806,7 @@ mod tests {
     fn jec_dep_info_keeps_only_logical_files_and_secret_digests() -> Result<(), Box<dyn std::error::Error>> {
         let environment = BTreeMap::from([("PRIVATE_VALUE".to_string(), PathBuf::from("do-not-retain-me"))]);
         let bytes = b"libfixture.rlib: incan-source/src/lib.rs incan-source/src/space\\ file.rs \\\n incan-source/src/dollar$$file.rs\n\n# env-dep:ABSENT\n# env-dep:PRIVATE_VALUE=do-not-retain-me\n";
-        let observed = super::parse_rustc_dep_info(bytes, &environment)?;
+        let observed = super::direct_compiler::parse_rustc_dep_info(bytes, &environment)?;
         assert_eq!(
             observed.files,
             [
@@ -7538,7 +5826,7 @@ mod tests {
         assert!(!format!("{observed:?}").contains("do-not-retain-me"));
 
         let mismatched = b"libfixture.rlib: incan-source/src/lib.rs\n# env-dep:PRIVATE_VALUE=different\n";
-        assert!(super::parse_rustc_dep_info(mismatched, &environment).is_err());
+        assert!(super::direct_compiler::parse_rustc_dep_info(mismatched, &environment).is_err());
         Ok(())
     }
 
@@ -7546,7 +5834,7 @@ mod tests {
     fn jec_compiler_evidence_hashes_the_selected_target_closure() -> Result<(), Box<dyn std::error::Error>> {
         let rustc = rustc_path()?;
         let target = rustc_host_target(&rustc)?;
-        let evidence = super::direct_rustc_compiler_evidence(&rustc, &target)?;
+        let evidence = super::direct_compiler::direct_rustc_compiler_evidence(&rustc, &target)?;
         assert_eq!(evidence.host(), target);
         assert_eq!(evidence.target(), target);
         assert_eq!(evidence.binary_digest(), digest_bytes(&fs::read(&rustc)?));
@@ -7609,8 +5897,13 @@ fi
         );
 
         let no_capacity = OvenStore::new(root.path().join("no-capacity"), OvenStoreLimits::new(0, 0, 0));
-        let unavailable = super::retain_direct_rustc_compiler(&no_capacity, &receipt("0.1.0")?, &rustc, "test-target")?;
-        let super::OvenDirectRustcCompilerRetention::Unavailable { reason } = unavailable else {
+        let unavailable = super::direct_compiler::retain_direct_rustc_compiler(
+            &no_capacity,
+            &receipt("0.1.0")?,
+            &rustc,
+            "test-target",
+        )?;
+        let super::direct_compiler::OvenDirectRustcCompilerRetention::Unavailable { reason } = unavailable else {
             return Err("zero-capacity Store unexpectedly retained a compiler owner".into());
         };
         assert!(
@@ -7618,8 +5911,8 @@ fi
             "{reason}"
         );
 
-        let super::OvenDirectRustcCompilerRetention::Retained(first) =
-            super::retain_direct_rustc_compiler(&store, &receipt("0.1.0")?, &rustc, "test-target")?
+        let super::direct_compiler::OvenDirectRustcCompilerRetention::Retained(first) =
+            super::direct_compiler::retain_direct_rustc_compiler(&store, &receipt("0.1.0")?, &rustc, "test-target")?
         else {
             return Err("compiler owner was not retained".into());
         };
@@ -7635,12 +5928,12 @@ fi
         // host/toolchain must neither be rescanned nor silently replace the already admitted closure for a new
         // project-version receipt.
         fs::write(&ambient_driver, b"different ambient driver bytes")?;
-        let changed_ambient = super::direct_rustc_compiler_evidence(&rustc, "test-target")?;
+        let changed_ambient = super::direct_compiler::direct_rustc_compiler_evidence(&rustc, "test-target")?;
         assert_ne!(changed_ambient.closure_digest(), first.evidence.closure_digest());
         fs::write(&rustc, "#!/bin/sh\nexit 91\n")?;
 
-        let super::OvenDirectRustcCompilerRetention::Retained(second) =
-            super::retain_direct_rustc_compiler(&store, &receipt("0.2.0")?, &rustc, "test-target")?
+        let super::direct_compiler::OvenDirectRustcCompilerRetention::Retained(second) =
+            super::direct_compiler::retain_direct_rustc_compiler(&store, &receipt("0.2.0")?, &rustc, "test-target")?
         else {
             return Err("compatible compiler owner was not reused".into());
         };
@@ -7671,7 +5964,7 @@ fi
         let sysroot = super::rustc_sysroot(&rustc)?.canonicalize()?;
         let source = "pub fn answer() -> u32 { 42 }\n";
         let source_digest = digest_bytes(source.as_bytes());
-        let compiler = super::OvenDirectRustcCompilerEvidence {
+        let compiler = super::direct_compiler::OvenDirectRustcCompilerEvidence {
             binary_digest: rustc_digest,
             closure_digest: digest_bytes(b"test compiler closure"),
             host: target.clone(),
@@ -7679,7 +5972,7 @@ fi
             sysroot,
             members: Vec::new(),
         };
-        let bindings = super::OvenDirectRustcJecBindings {
+        let bindings = super::direct_compiler::OvenDirectRustcJecBindings {
             logical_source_root: "incan-source".to_string(),
             logical_entrypoint: "src/lib.rs".to_string(),
             externs: Vec::new(),
@@ -7695,7 +5988,7 @@ fi
             fs::create_dir_all(source_root.join("src"))?;
             fs::create_dir(&artifact_root)?;
             fs::write(source_root.join("src/lib.rs"), source)?;
-            let prepared = super::OvenPreparedDirectRustcLibrary {
+            let prepared = super::direct_compiler::OvenPreparedDirectRustcLibrary {
                 rustc: rustc.clone(),
                 compiler: Some(compiler.clone()),
                 compiler_owner: None,
@@ -7734,13 +6027,13 @@ fi
                 edition: "2021".to_string(),
                 features: Vec::new(),
                 frozen_environment: Some(
-                    super::freeze_direct_rustc_environment(std::env::vars_os(), &BTreeMap::new())
+                    super::direct_compiler::freeze_direct_rustc_environment(std::env::vars_os(), &BTreeMap::new())
                         .ok_or("test environment was not representable")?,
                 ),
             };
             let compiled = prepared.bind(&output, &bindings)?.compile()?;
             outputs.push(fs::read(&compiled.bake.output)?);
-            let super::OvenRustcDepInfoOutcome::Observed(observation) = compiled.observation else {
+            let super::direct_compiler::OvenRustcDepInfoOutcome::Observed(observation) = compiled.observation else {
                 return Err("relocated Rustc compilation did not provide dep-info".into());
             };
             observations.push(observation.files);
@@ -7783,7 +6076,7 @@ fi
         }
         let source = "pub fn answer() -> u32 { fixture_dep::answer() }\n";
         let source_digest = digest_bytes(source.as_bytes());
-        let compiler = super::OvenDirectRustcCompilerEvidence {
+        let compiler = super::direct_compiler::OvenDirectRustcCompilerEvidence {
             binary_digest: digest_bytes(&fs::read(&rustc)?),
             closure_digest: digest_bytes(b"test compiler closure"),
             host: target.clone(),
@@ -7791,7 +6084,7 @@ fi
             sysroot: super::rustc_sysroot(&rustc)?.canonicalize()?,
             members: Vec::new(),
         };
-        let bindings = super::OvenDirectRustcJecBindings {
+        let bindings = super::direct_compiler::OvenDirectRustcJecBindings {
             logical_source_root: "incan-source".to_string(),
             logical_entrypoint: "src/lib.rs".to_string(),
             externs: vec!["extern/0000".to_string()],
@@ -7811,7 +6104,7 @@ fi
             fs::copy(&dependency_output, &relocated_dependency)?;
             let dependency_root = dependency_root.canonicalize()?;
             let relocated_dependency = relocated_dependency.canonicalize()?;
-            let prepared = super::OvenPreparedDirectRustcLibrary {
+            let prepared = super::direct_compiler::OvenPreparedDirectRustcLibrary {
                 rustc: rustc.clone(),
                 compiler: Some(compiler.clone()),
                 compiler_owner: None,
@@ -7850,7 +6143,7 @@ fi
                 edition: "2021".to_string(),
                 features: Vec::new(),
                 frozen_environment: Some(
-                    super::freeze_direct_rustc_environment(std::env::vars_os(), &BTreeMap::new())
+                    super::direct_compiler::freeze_direct_rustc_environment(std::env::vars_os(), &BTreeMap::new())
                         .ok_or("test environment was not representable")?,
                 ),
             };
@@ -7990,7 +6283,7 @@ fi
         let debug = native_input_fixture("debug")?;
         let release = native_input_fixture("release")?;
         for fixture in [&debug, &release] {
-            let view = super::OvenNativeInputView::from_store_payload(&fixture.owner)?;
+            let view = super::native_input::OvenNativeInputView::from_store_payload(&fixture.owner)?;
             assert_eq!(view.identity(), fixture.owner.manifest.identity);
             assert_eq!(view.receipt_identity(), Some(fixture.receipt.identity.as_str()));
             assert_eq!(view.build_unit_identity(), fixture.receipt.build_unit_identity);
@@ -8020,7 +6313,7 @@ fi
     fn native_input_view_attaches_selected_alias_and_preserves_role_closure() -> Result<(), Box<dyn std::error::Error>>
     {
         let fixture = native_input_fixture("debug")?;
-        let view = super::OvenNativeInputView::from_store_payload(&fixture.owner)?;
+        let view = super::native_input::OvenNativeInputView::from_store_payload(&fixture.owner)?;
         let before = super::trusted_artifact_plan_for_source_evidence(&view.plan, &view.artifacts, "generated-root")?;
         let chosen = view
             .registry_candidates("generated-root")?
@@ -8058,8 +6351,8 @@ fi
     {
         let first = native_input_fixture("debug")?;
         let second = native_input_fixture("debug")?;
-        let view = super::OvenNativeInputView::from_store_payload(&first.owner)?;
-        let other = super::OvenNativeInputView::from_store_payload(&second.owner)?;
+        let view = super::native_input::OvenNativeInputView::from_store_payload(&first.owner)?;
+        let other = super::native_input::OvenNativeInputView::from_store_payload(&second.owner)?;
         let foreign = other
             .registry_candidates("generated-root")?
             .pop()
@@ -8098,7 +6391,7 @@ fi
     fn compiler_runtime_member_roundtrip_preserves_original_role_and_rechecks_bytes()
     -> Result<(), Box<dyn std::error::Error>> {
         let fixture = native_input_fixture("debug")?;
-        let view = super::OvenNativeInputView::from_store_payload(&fixture.owner)?;
+        let view = super::native_input::OvenNativeInputView::from_store_payload(&fixture.owner)?;
         let members = view.named_candidates("generated-root")?;
         let member = members.first().ok_or("original runtime member absent")?;
         let request = serde_json::json!({
@@ -8145,10 +6438,11 @@ fi
                 .any(|(name, _)| name == "private")
         );
         assert!(view.attach_for_source("compiler-helper", &selected).is_err());
-        fs::write(
-            fixture.owner.artifact_root.join("release/deps/libruntime.rlib"),
-            b"changed named runtime",
-        )?;
+        // Admission seals materialized files read-only; replace this test-owned entry rather than writing
+        // through the seal.
+        let runtime_member = fixture.owner.artifact_root.join("release/deps/libruntime.rlib");
+        fs::remove_file(&runtime_member)?;
+        fs::write(&runtime_member, b"changed named runtime")?;
         assert!(matches!(
             view.attach_for_source("generated-root", &selected),
             Err(OvenRustcError::ArtifactDigestMismatch { .. })
@@ -8160,9 +6454,12 @@ fi
     fn native_input_view_queries_are_read_free_but_attachment_rechecks_selected_bytes()
     -> Result<(), Box<dyn std::error::Error>> {
         let fixture = native_input_fixture("debug")?;
-        let view = super::OvenNativeInputView::from_store_payload(&fixture.owner)?;
+        let view = super::native_input::OvenNativeInputView::from_store_payload(&fixture.owner)?;
         let original_plan = view.plan.clone();
+        // Admission seals materialized files read-only, so replacing this test-owned entry models the external
+        // mutation under test rather than writing through the seal.
         let file = fixture.owner.artifact_root.join("release/deps/libserde_fixture.rlib");
+        fs::remove_file(&file)?;
         fs::write(&file, b"changed after admission")?;
         let chosen = view
             .registry_candidates("generated-root")?
@@ -8188,18 +6485,18 @@ fi
         assert_ne!(first.owner.artifact_root, second.owner.artifact_root);
         let original_payload = first.owner.payload.clone();
         first.owner.payload.push(b' ');
-        assert!(super::OvenNativeInputView::from_store_payload(&first.owner).is_err());
+        assert!(super::native_input::OvenNativeInputView::from_store_payload(&first.owner).is_err());
         first.owner.payload = original_payload;
         let original_manifest = first.owner.manifest.clone();
         first.owner.manifest.intent.profile = "release".to_string();
-        assert!(super::OvenNativeInputView::from_store_payload(&first.owner).is_err());
+        assert!(super::native_input::OvenNativeInputView::from_store_payload(&first.owner).is_err());
         first.owner.manifest = original_manifest;
-        assert!(super::OvenNativeInputView::from_store_payload(&first.owner).is_ok());
+        assert!(super::native_input::OvenNativeInputView::from_store_payload(&first.owner).is_ok());
         first.owner.manifest = second.owner.manifest.clone();
         first.owner.artifact_root = second.owner.artifact_root.clone();
         first.owner.payload = second.owner.payload.clone();
         assert!(
-            super::OvenNativeInputView::from_store_payload(&first.owner).is_err(),
+            super::native_input::OvenNativeInputView::from_store_payload(&first.owner).is_err(),
             "even identical valid records cannot retarget the original selected lease to another root"
         );
         Ok(())
@@ -8212,7 +6509,7 @@ fi
             fixture.root.path().join("store"),
             OvenStoreLimits::new(1, 4 * 1024 * 1024, 4 * 1024 * 1024),
         );
-        let view = super::OvenNativeInputView::from_store_payload(&fixture.owner)?;
+        let view = super::native_input::OvenNativeInputView::from_store_payload(&fixture.owner)?;
         let attached = view.attach_for_source("generated-root", &[])?;
         bounded.prune()?;
         let held = fixture.store.inspect()?;
@@ -8231,7 +6528,7 @@ fi
     #[test]
     fn native_input_view_refuses_selected_symlink_after_admission() -> Result<(), Box<dyn std::error::Error>> {
         let fixture = native_input_fixture("debug")?;
-        let view = super::OvenNativeInputView::from_store_payload(&fixture.owner)?;
+        let view = super::native_input::OvenNativeInputView::from_store_payload(&fixture.owner)?;
         let selected = fixture.owner.artifact_root.join("release/deps/libserde_fixture.rlib");
         let foreign = fixture.root.path().join("foreign.rlib");
         fs::write(&foreign, b"registry bytes")?;
@@ -12016,13 +10313,23 @@ fi
                 .contains("inconsistent identity, kind, or base evidence")
         );
 
+        // The dependency's path has to exist: digesting a path dependency reads the directory, so a bare
+        // placeholder makes this case fail on "provider artifact root is not a directory" long before reaching
+        // the source-evidence refusal it exists to check.
+        let legacy_path_dependency = project.path().join("legacy-path-dependency");
+        std::fs::create_dir_all(legacy_path_dependency.join("src"))?;
+        std::fs::write(
+            legacy_path_dependency.join("Cargo.toml"),
+            "[package]\nname = \"dev_fixture\"\nversion = \"0.1.0\"\nedition = \"2021\"\n",
+        )?;
+        std::fs::write(legacy_path_dependency.join("src/lib.rs"), "")?;
         let path_dependency = DependencySpec {
             crate_name: "dev_fixture".to_string(),
             version: Some("0.1.0".to_string()),
             features: vec!["test-support".to_string()],
             default_features: false,
             source: DependencySource::Path {
-                path: PathBuf::from("legacy-path-dependency"),
+                path: legacy_path_dependency.clone(),
             },
             optional: false,
             package: None,
@@ -12038,15 +10345,20 @@ fi
                     dependency_digest: "sha256:legacy-path-envelope".to_string(),
                 },
             );
-        let Err(error) = project_inspection_test_dependency_envelope_supports_dependencies(
-            &payload,
-            std::slice::from_ref(&path_dependency),
-        ) else {
-            return Err(
-                "persisted Cargo path envelope authorized execution without checked Oven source evidence".into(),
-            );
-        };
-        assert!(error.to_string().contains("has no checked Oven source-unit identity"));
+        // A persisted Cargo path root whose digest does not match the dependency is not supported, so the
+        // envelope cannot authorize it.
+        //
+        // This substrate carries the shapes, not the decision. On `work/oven-hot-path` the same case is a hard
+        // refusal naming "no checked Oven source-unit identity"; producing that error is a Gate 6 blocker, and
+        // the blocker that adds it replaces this with the negative case. Asserting the refusal here would be
+        // asserting behaviour this branch does not bring.
+        assert!(
+            !project_inspection_test_dependency_envelope_supports_dependencies(
+                &payload,
+                std::slice::from_ref(&path_dependency),
+            )?,
+            "an unmatched persisted Cargo path root must not be reported as supported"
+        );
 
         payload
             .test_dependency_envelope
@@ -12290,6 +10602,7 @@ pub(crate) fn incan_owned_cargo() -> Option<PathBuf> {
 /// each child output digest into its parent output identity. Git, optional/feature-driven roots, build scripts, and
 /// unsealed registry closures remain explicit unsupported inputs.
 #[cfg(test)]
+#[allow(dead_code, reason = "Gate 6 of RFC 119 is the reader; see the note above")]
 pub(crate) fn materialize_declared_rust_libraries(
     output_root: &Path,
     rustc: &Path,
